Separated the icons into individual files and imported them correctly into the views.
