+++ src/app/deliveries/page.tsx
@@ -1,5 +1,6 @@
 import { poppins400, poppins600, poppins700 } from "@/commons/fonts";
-import { ChevronArrowDown, LeftArrowIcon } from "@/commons/icons";
+import { ChevronArrowDown } from "@/commons/icons/ChevronArrowDown";
+import { LeftArrowIcon } from "@/commons/icons/LeftArrowIcon";
 import { Navbar } from "@/components/Navbar";
 import { Progress } from "antd";
 import Image from "next/image";
