+++ src/app/get-packages/page.tsx
@@ -1,4 +1,7 @@
-import { CheckboxCheck, CheckboxEmpty, LeftArrowIcon } from '@/commons/icons';
+// import { CheckboxCheck, CheckboxEmpty, LeftArrowIcon } from '@/commons/icons';
+import { CheckboxCheck } from '@/commons/icons/CheckboxCheck';
+import { CheckboxEmpty } from '@/commons/icons/CheckboxEmpty';
+import { LeftArrowIcon } from '@/commons/icons/LeftArrowIcon';
 import { Navbar } from '@/components/Navbar';
 import { poppins400, poppins700 } from '@/commons/fonts';
