+++ src/app/manage-orders/page.tsx
@@ -7,7 +7,8 @@ import adminLogo from "../../../public/adminFoto.png";
 import delivery1 from "../../../public/delivery1.png";
 import delivery2 from "../../../public/delivery2.png";
 import { poppins300, poppins400, poppins500, poppins600, poppins700 } from "@/commons/fonts";
-import { TriangleDownArrow, PlusIcon } from "@/commons/icons";
+import { TriangleDownArrow } from "@/commons/icons/TriangleDownArrow";
+import { PlusIcon } from "@/commons/icons/PlusIcon";
 import { Progress } from "antd";
 import { DateTime } from "luxon";
 import { fiveDays, injectTime, getPreviousFiveDays, getNextFiveDays } from "@/utils/calendar";
