+++ src/app/working-day/page.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import { Navbar } from "@/components/Navbar";
-import { TriangleDownArrow } from "@/commons/icons";
-import { TrashIcon } from "@/commons/icons";
+import { TriangleDownArrow } from "@/commons/icons/TriangleDownArrow";
+import { TrashIcon } from "@/commons/icons/TrashIcon";
 import box from "../../../public/Box.png";
 import { poppins300, poppins400, poppins500, poppins600, poppins700, poppins800 } from "@/commons/fonts";
