+++ src/commons/icons/LockIcon.tsx
@@ -1,0 +1,19 @@
+export function LockIcon({ className = "" }) {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        strokeWidth="2"
+        stroke="currentColor"
+        fill="none"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        className={className}
+      >
+        <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+        <path d="M5 13a2 2 0 0 1 2 -2h10a2 2 0 0 1 2 2v6a2 2 0 0 1 -2 2h-10a2 2 0 0 1 -2 -2v-6z" />
+        <path d="M11 16a1 1 0 1 0 2 0a1 1 0 0 0 -2 0" />
+        <path d="M8 11v-4a4 4 0 1 1 8 0v4" />
+      </svg>
+    );
+  }
