+++ src/commons/icons/TriangleDownArrow.tsx
@@ -1,0 +1,12 @@
+export function TriangleDownArrow({ className = "", onClick }: { className?: string; onClick?: () => void }) {
+    return (
+      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor" fill="none" strokeLinecap="round" strokeLinejoin="round" className={className} onClick={onClick}>
+        <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+        <path
+          d="M18 9c.852 0 1.297 .986 .783 1.623l-.076 .084l-6 6a1 1 0 0 1 -1.32 .083l-.094 -.083l-6 -6l-.083 -.094l-.054 -.077l-.054 -.096l-.017 -.036l-.027 -.067l-.032 -.108l-.01 -.053l-.01 -.06l-.004 -.057v-.118l.005 -.058l.009 -.06l.01 -.052l.032 -.108l.027 -.067l.07 -.132l.065 -.09l.073 -.081l.094 -.083l.077 -.054l.096 -.054l.036 -.017l.067 -.027l.108 -.032l.053 -.01l.06 -.01l.057 -.004l12.059 -.002z"
+          strokeWidth="0"
+          fill="currentColor"
+        />
+      </svg>
+    );
+  }
